testata funzione conta Proprietari con automobile immatricolata dopo un certo anno

+++ src/it/proprietarioautomobile/dao/proprietario/ProprietarioDAO.java
@@ -4,5 +4,5 @@ import it.proprietarioautomobile.dao.IBaseDAO;
 import it.proprietarioautomobile.model.Proprietario;
 
 public interface ProprietarioDAO extends IBaseDAO<Proprietario> {
-	public int contaProprietariConAutomobiliImmatricolateDopo(int anno) throws Exception;
+	public long contaProprietariConAutomobiliImmatricolateDopo(int anno) throws Exception;
 }
